test(mitm-addon): split built-in registry cache tests
