--- crates/runner/mitm-addon/tests/test_registry_builtin_cache.py
@@ -1,1479 +1,0 @@
-"""Tests for registry built-in firewall resolution and core-cache behavior."""
-
-import json
-import os
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-import builtin_firewall_cache
-import builtin_host_policy
-import matching
-import registry
-import registry_firewalls
-from tests.registry_helpers import (
-    assert_invalid_builtin_vm,
-    builtin_vm,
-    inline_vm,
-    write_multi_vm_registry,
-    write_trusted_catalog_cache_text,
-)
-
-
-def _first_firewall_core(compiled_firewalls: matching.CompiledFirewallSet):
-    return compiled_firewalls.firewalls[0].core
-
-
-def _cache_firewall(name: str, base: str, *, rules: list[str] | None = None) -> dict:
-    return {
-        "name": name,
-        "apis": [
-            {
-                "base": base,
-                "hostPolicy": {
-                    "kind": "providerOwned",
-                    "exactHosts": [base.removeprefix("https://")],
-                },
-                "auth": {
-                    "awsSigv4": {
-                        "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                        "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                    }
-                },
-                "permissions": [{"name": "read", "rules": rules or ["GET /items"]}],
-            }
-        ],
-    }
-
-
-def _github_cache_firewall() -> dict:
-    return {
-        "name": "github",
-        "apis": [
-            {
-                "base": "https://api.github.com",
-                "auth": {"headers": {}},
-                "permissions": [{"name": "read", "rules": ["GET /repos/{owner}/{repo}"]}],
-            }
-        ],
-    }
-
-
-def _zendesk_cache_firewall() -> dict:
-    return {
-        "name": "zendesk",
-        "apis": [
-            {
-                "base": "https://${{ vars.ZENDESK_SUBDOMAIN }}.zendesk.com",
-                "auth": {"headers": {}},
-                "permissions": [{"name": "read", "rules": ["GET /api/v2/tickets"]}],
-            }
-        ],
-    }
-
-
-def _write_catalog_cache(
-    path,
-    *,
-    digest: str,
-    version: str,
-    firewalls: dict[str, dict],
-) -> None:
-    write_trusted_catalog_cache_text(
-        path,
-        json.dumps(
-            {
-                "schemaVersion": 1,
-                "catalogDigest": digest,
-                "catalogVersion": version,
-                "updatedAt": "2026-07-07T00:00:00.000Z",
-                "firewalls": firewalls,
-            },
-            sort_keys=True,
-        ),
-    )
-
-
-def _write_registry_with_cache(
-    tmp_path,
-    vms: dict,
-    firewalls: dict[str, dict],
-    *,
-    digest: str = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-    version: str = "catalog-a",
-):
-    registry_path = tmp_path / "registry.json"
-    cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-    write_multi_vm_registry(registry_path, vms)
-    _write_catalog_cache(
-        cache_path,
-        digest=digest,
-        version=version,
-        firewalls=firewalls,
-    )
-    return registry_path, cache_path
-
-
-def _catalog_snapshot(
-    *,
-    digest_char: str,
-    version: str,
-    firewalls: dict[str, dict],
-) -> builtin_firewall_cache.BuiltinFirewallCatalogSnapshot:
-    key = builtin_firewall_cache.CatalogFileKey(
-        absolute_path=f"catalog-cache/{version}.json",
-        st_dev=1,
-        st_ino=1,
-        st_mtime_ns=len(version),
-        st_size=len(firewalls),
-    )
-    digest = f"sha256:{digest_char * 64}"
-    return builtin_firewall_cache.BuiltinFirewallCatalogSnapshot(
-        dependency_file_key=key,
-        catalog=builtin_firewall_cache.BuiltinFirewallCatalog(
-            identity=builtin_firewall_cache.CatalogIdentity(
-                source="cache",
-                catalog_digest=digest,
-                catalog_version=version,
-                file_key=key,
-            ),
-            firewalls=firewalls,
-        ),
-        cache_path=key.absolute_path,
-    )
-
-
-def _assert_invalid_builtin_vm_with_cache(
-    *,
-    registry_path,
-    cache_path,
-    mitm_ctx,
-    expected_message: str,
-) -> None:
-    with mitm_ctx(
-        registry_path=str(registry_path),
-        builtin_firewall_catalog_cache_path=str(cache_path),
-    ):
-        invalid_vm = assert_invalid_builtin_vm(registry_path)
-
-    assert expected_message in invalid_vm.message
-
-
-def _assert_cache_firewall_is_invalid(
-    tmp_path,
-    mitm_ctx,
-    firewall: dict,
-    *,
-    expected_message: str = "catalog cache unavailable: cache_invalid",
-    cache_mode: int | None = None,
-) -> None:
-    registry_path = tmp_path / "registry.json"
-    cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-    _write_catalog_cache(
-        cache_path,
-        digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-        version="catalog-a",
-        firewalls={"fallback": firewall},
-    )
-    if cache_mode is not None:
-        cache_path.chmod(cache_mode)
-    write_multi_vm_registry(
-        registry_path,
-        {"10.200.0.1": builtin_vm("run-fallback", "fallback")},
-    )
-
-    _assert_invalid_builtin_vm_with_cache(
-        registry_path=registry_path,
-        cache_path=cache_path,
-        mitm_ctx=mitm_ctx,
-        expected_message=expected_message,
-    )
-
-
-class TestRegistryBuiltinCache:
-    def test_registry_firewalls_does_not_expose_runtime_bundled_catalog(self):
-        assert not hasattr(registry_firewalls, "BUILTIN_FIREWALLS")
-
-    def test_resolved_firewall_entries_enforces_cache_key_alignment(self):
-        inline_firewall = {"name": "inline", "apis": []}
-
-        registry_firewalls.ResolvedFirewallEntries(None, None)
-        registry_firewalls.ResolvedFirewallEntries([], ())
-        registry_firewalls.ResolvedFirewallEntries([inline_firewall], (None,))
-
-        with pytest.raises(ValueError, match="absent when firewalls are absent"):
-            registry_firewalls.ResolvedFirewallEntries(None, (None,))
-        with pytest.raises(ValueError, match="present when firewalls are present"):
-            registry_firewalls.ResolvedFirewallEntries([inline_firewall], None)
-        with pytest.raises(ValueError, match="align with resolved firewalls"):
-            registry_firewalls.ResolvedFirewallEntries([inline_firewall], ())
-
-    def test_builtin_firewall_entry_resolves_from_catalog_cache(self, tmp_path, mitm_ctx):
-        registry_path, cache_path = _write_registry_with_cache(
-            tmp_path,
-            {"10.200.0.1": builtin_vm("run-github", "github")},
-            {"github": _github_cache_firewall()},
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["name"] == "github"
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == "https://api.github.com"
-        assert vm_info["firewalls"][0]["apis"][0]["id"] == "run-github:0"
-
-    def test_builtin_firewall_entry_prefers_runner_catalog_cache(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-cache-only", "cache-only")},
-        )
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={
-                "cache-only": _cache_firewall(
-                    "cache-only",
-                    "https://cache-only.example.com",
-                )
-            },
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        api = vm_info["firewalls"][0]["apis"][0]
-        assert api["base"] == "https://cache-only.example.com"
-        assert api["auth"]["awsSigv4"]["accessKeyId"] == "${{ secrets.AWS_ACCESS_KEY_ID }}"
-        assert api["hostPolicy"]["exactHosts"] == ["cache-only.example.com"]
-        assert isinstance(
-            api[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER],
-            builtin_host_policy.CompiledBuiltinHostPolicy,
-        )
-
-    def test_catalog_api_ids_are_reassigned_per_vm(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        firewall = _cache_firewall("cache-only", "https://cache-only.example.com")
-        firewall["apis"][0]["id"] = "catalog-owned-id"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-cache-only", "cache-only")},
-        )
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"cache-only": firewall},
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["id"] == "run-cache-only:0"
-
-    def test_catalog_identity_is_checked_only_for_cached_builtin_registry(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={
-                "cache-only": _cache_firewall(
-                    "cache-only",
-                    "https://cache.example.com",
-                )
-            },
-        )
-        original_catalog_file_key = registry_firewalls.catalog_file_key
-
-        with (
-            mitm_ctx(
-                registry_path=str(registry_path),
-                builtin_firewall_catalog_cache_path=str(cache_path),
-            ),
-            patch.object(
-                registry_firewalls,
-                "catalog_file_key",
-                wraps=original_catalog_file_key,
-            ) as catalog_file_key,
-        ):
-            missing_state = registry.load_registry_state(str(registry_path))
-            assert isinstance(missing_state, registry.RegistryUnavailable)
-            assert missing_state.reason == "stat_failed"
-            assert catalog_file_key.call_count == 0
-
-            write_multi_vm_registry(
-                registry_path,
-                {"10.200.0.1": inline_vm("run-inline")},
-            )
-            os.utime(
-                registry_path,
-                ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000),
-            )
-            inline_state = registry.load_registry_state(str(registry_path))
-            cached_inline_state = registry.load_registry_state(str(registry_path))
-
-            assert not isinstance(inline_state, registry.RegistryUnavailable)
-            assert cached_inline_state is inline_state
-            assert inline_state.vms["10.200.0.1"]["runId"] == "run-inline"
-            assert catalog_file_key.call_count == 0
-
-            write_multi_vm_registry(
-                registry_path,
-                {"10.200.0.1": builtin_vm("run-cache-only", "cache-only")},
-            )
-            os.utime(
-                registry_path,
-                ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001),
-            )
-            builtin_state = registry.load_registry_state(str(registry_path))
-
-            assert not isinstance(builtin_state, registry.RegistryUnavailable)
-            assert (
-                builtin_state.vms["10.200.0.1"]["firewalls"][0]["apis"][0]["base"]
-                == "https://cache.example.com"
-            )
-            assert catalog_file_key.call_count == 0
-
-            cached_builtin_state = registry.load_registry_state(str(registry_path))
-
-        assert cached_builtin_state is builtin_state
-        assert catalog_file_key.call_count == 1
-
-    def test_malformed_registry_cache_is_independent_of_catalog_identity(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        registry_path.write_text("{ broken")
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"unused": _cache_firewall("unused", "https://unused-a.example.com")},
-        )
-        os.utime(
-            cache_path,
-            ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000),
-        )
-        original_catalog_file_key = registry_firewalls.catalog_file_key
-
-        with (
-            mitm_ctx(
-                registry_path=str(registry_path),
-                builtin_firewall_catalog_cache_path=str(cache_path),
-            ) as log,
-            patch.object(
-                registry_firewalls,
-                "catalog_file_key",
-                wraps=original_catalog_file_key,
-            ) as catalog_file_key,
-            patch.object(registry.json, "loads", wraps=registry.json.loads) as json_loads,
-        ):
-            first_state = registry.load_registry_state(str(registry_path))
-            _write_catalog_cache(
-                cache_path,
-                digest="sha256:bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                version="catalog-b",
-                firewalls={"unused": _cache_firewall("unused", "https://unused-b.example.com")},
-            )
-            os.utime(
-                cache_path,
-                ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001),
-            )
-            second_state = registry.load_registry_state(str(registry_path))
-
-        assert isinstance(first_state, registry.RegistryUnavailable)
-        assert second_state is first_state
-        assert json_loads.call_count == 1
-        assert catalog_file_key.call_count == 0
-        assert log.warn.call_count == 1
-
-    def test_registry_snapshot_uses_actual_loaded_catalog_file_key(
-        self, tmp_path, monkeypatch, mitm_ctx
-    ):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-racy-cache", "racy-cache")},
-        )
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={
-                "racy-cache": _cache_firewall(
-                    "racy-cache",
-                    "https://cache.example.com",
-                )
-            },
-        )
-        original_catalog_file_key = registry_firewalls.catalog_file_key
-        calls = 0
-
-        def racy_catalog_file_key(cache_path: str | None):
-            nonlocal calls
-            calls += 1
-            if calls == 1:
-                return None
-            return original_catalog_file_key(cache_path)
-
-        monkeypatch.setattr(registry_firewalls, "catalog_file_key", racy_catalog_file_key)
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            assert context is not None
-            vm_info, compiled_firewalls, _ = context
-            assert compiled_firewalls is not None
-            assert vm_info["firewalls"][0]["apis"][0]["base"] == "https://cache.example.com"
-
-            cache_path.unlink()
-
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            state = registry.load_registry_state(str(registry_path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-        assert (
-            "catalog cache unavailable: cache_file_missing"
-            in state.invalid_vms["10.200.0.1"].message
-        )
-
-    def test_builtin_registry_reload_when_catalog_cache_appears(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-late-cache", "late-cache")},
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            _write_catalog_cache(
-                cache_path,
-                digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                version="catalog-a",
-                firewalls={
-                    "late-cache": _cache_firewall(
-                        "late-cache",
-                        "https://cache.example.com",
-                    )
-                },
-            )
-            second_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert first_context is None
-        assert second_context is not None
-        second_vm_info, second_compiled, _ = second_context
-        assert second_compiled is not None
-        assert second_vm_info["firewalls"][0]["apis"][0]["base"] == "https://cache.example.com"
-
-    def test_unknown_builtin_registry_reload_when_catalog_cache_appears(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-cache-only", "cache-only")},
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            first_state = registry.load_registry_state(str(registry_path))
-            _write_catalog_cache(
-                cache_path,
-                digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                version="catalog-a",
-                firewalls={
-                    "cache-only": _cache_firewall(
-                        "cache-only",
-                        "https://cache.example.com",
-                    )
-                },
-            )
-            second_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert first_context is None
-        assert not isinstance(first_state, registry.RegistryUnavailable)
-        assert first_state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-        assert second_context is not None
-        second_vm_info, second_compiled, _ = second_context
-        assert second_compiled is not None
-        assert second_vm_info["firewalls"][0]["apis"][0]["base"] == "https://cache.example.com"
-
-    def test_missing_runner_catalog_cache_fails_closed(self, tmp_path):
-        cache_path = tmp_path / "missing-builtin-firewall-catalog-cache.json"
-        snapshot = builtin_firewall_cache.BuiltinFirewallCatalogSnapshot(
-            dependency_file_key=None,
-            catalog=None,
-            cache_path=str(cache_path),
-            unavailable_reason="cache_file_missing",
-        )
-
-        with pytest.raises(
-            registry_firewalls.FirewallEntryResolutionError,
-            match='builtin firewall "fallback" catalog cache unavailable: '
-            f"cache_file_missing \\({cache_path}\\)",
-        ):
-            registry_firewalls.resolve_firewall_entries(
-                builtin_vm("run-fallback", "fallback"),
-                builtin_firewall_catalog_snapshot=snapshot,
-            )
-
-    def test_runner_catalog_cache_missing_firewall_fails_closed(self):
-        snapshot = _catalog_snapshot(
-            digest_char="a",
-            version="catalog-a",
-            firewalls={
-                "other": _cache_firewall(
-                    "other",
-                    "https://cache.example.com",
-                )
-            },
-        )
-        assert snapshot.catalog is not None
-        rewritten_key = builtin_firewall_cache.CatalogFileKey(
-            absolute_path="catalog-cache/catalog-a.json",
-            st_dev=1,
-            st_ino=2,
-            st_mtime_ns=3,
-            st_size=4,
-        )
-        rewritten_snapshot = builtin_firewall_cache.BuiltinFirewallCatalogSnapshot(
-            dependency_file_key=rewritten_key,
-            catalog=builtin_firewall_cache.BuiltinFirewallCatalog(
-                identity=builtin_firewall_cache.CatalogIdentity(
-                    source=snapshot.catalog.identity.source,
-                    catalog_digest=snapshot.catalog.identity.catalog_digest,
-                    catalog_version=snapshot.catalog.identity.catalog_version,
-                    file_key=rewritten_key,
-                ),
-                firewalls=snapshot.catalog.firewalls,
-            ),
-            cache_path=rewritten_key.absolute_path,
-        )
-        expected_digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-
-        for current_snapshot in (snapshot, rewritten_snapshot):
-            with pytest.raises(
-                registry_firewalls.FirewallEntryResolutionError,
-                match='builtin firewall "fallback" missing from catalog cache '
-                rf"\(catalog_digest={expected_digest}, catalog_version=catalog-a\)",
-            ):
-                registry_firewalls.resolve_firewall_entries(
-                    builtin_vm("run-fallback", "fallback"),
-                    builtin_firewall_catalog_snapshot=current_snapshot,
-                )
-
-    def test_registry_reload_pins_one_catalog_snapshot_for_builtin_entries(
-        self, tmp_path, monkeypatch, mitm_ctx
-    ):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {
-                "10.200.0.1": {
-                    "runId": "run-pinned-cache",
-                    "firewalls": [
-                        {"kind": "builtin", "name": "alpha"},
-                        {"kind": "builtin", "name": "beta"},
-                    ],
-                }
-            },
-        )
-        snapshots = [
-            _catalog_snapshot(
-                digest_char="a",
-                version="catalog-a",
-                firewalls={
-                    "alpha": _cache_firewall("alpha", "https://alpha-a.example.com"),
-                    "beta": _cache_firewall("beta", "https://beta-a.example.com"),
-                },
-            ),
-            _catalog_snapshot(
-                digest_char="b",
-                version="catalog-b",
-                firewalls={
-                    "alpha": _cache_firewall("alpha", "https://alpha-b.example.com"),
-                    "beta": _cache_firewall("beta", "https://beta-b.example.com"),
-                },
-            ),
-        ]
-        load_calls: list[str | None] = []
-
-        def load_catalog_snapshot(
-            cache_path: str | None,
-        ) -> builtin_firewall_cache.BuiltinFirewallCatalogSnapshot:
-            load_calls.append(cache_path)
-            return snapshots[min(len(load_calls) - 1, len(snapshots) - 1)]
-
-        monkeypatch.setattr(
-            registry_firewalls,
-            "load_catalog_snapshot",
-            load_catalog_snapshot,
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        bases = [api["base"] for firewall in vm_info["firewalls"] for api in firewall["apis"]]
-        assert bases == ["https://alpha-a.example.com", "https://beta-a.example.com"]
-        assert load_calls == [str(cache_path)]
-
-    def test_runner_catalog_cache_accepts_valid_template_base(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {
-                "10.200.0.1": builtin_vm(
-                    "run-template",
-                    "templated",
-                    {"TENANT": "acme"},
-                )
-            },
-        )
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={
-                "templated": {
-                    "name": "templated",
-                    "apis": [
-                        {
-                            "base": "https://${{ vars.TENANT }}.example.com",
-                            "auth": {"headers": {}},
-                            "permissions": [{"name": "read", "rules": ["GET /items"]}],
-                        }
-                    ],
-                }
-            },
-        )
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == "https://acme.example.com"
-
-    def test_malformed_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_trusted_catalog_cache_text(cache_path, '{"schemaVersion":1}')
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-fallback", "fallback")},
-        )
-
-        _assert_invalid_builtin_vm_with_cache(
-            registry_path=registry_path,
-            cache_path=cache_path,
-            mitm_ctx=mitm_ctx,
-            expected_message="catalog cache unavailable: cache_invalid",
-        )
-
-    def test_empty_api_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"fallback": {"name": "fallback", "apis": []}},
-        )
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-fallback", "fallback")},
-        )
-
-        _assert_invalid_builtin_vm_with_cache(
-            registry_path=registry_path,
-            cache_path=cache_path,
-            mitm_ctx=mitm_ctx,
-            expected_message="catalog cache unavailable: cache_invalid",
-        )
-
-    def test_malformed_static_base_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["base"] = "not-a-url"
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    @pytest.mark.parametrize(
-        "base",
-        [
-            "https://api.example.com/a/../admin",
-            "https://api.example.com/%252e%252e/admin",
-            "https://api.example.com/..;version=1/admin",
-            "https://api.example.com/%255cadmin",
-            "https://api.example.com/v1/../{org}",
-            "https://${{ vars.TENANT }}.example.com/v1/../items",
-            "${{ vars.API_BASE_URL }}/v1/%2e%2e/items",
-        ],
-    )
-    def test_unsafe_base_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx, base):
-        firewall = _cache_firewall("fallback", base)
-        firewall["apis"][0].pop("hostPolicy")
-
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_parameterized_base_runner_catalog_cache_fails_closed(
-        self, tmp_path, mitm_ctx
-    ):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["base"] = "https://api.{tenant+}.example.com"
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_template_base_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["base"] = "https://${{ secrets.TENANT }}.example.com"
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_template_parameter_base_runner_catalog_cache_fails_closed(
-        self, tmp_path, mitm_ctx
-    ):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["base"] = "https://${{ vars.TENANT }}.{tenant+}.example.com"
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_non_ascii_template_variable_runner_catalog_cache_fails_closed(
-        self, tmp_path, mitm_ctx
-    ):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["base"] = "https://${{ vars.\u00e9 }}.example.com"
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_auth_base_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["auth"] = {"base": "http://auth.example.com"}
-
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_world_writable_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-
-        _assert_cache_firewall_is_invalid(
-            tmp_path,
-            mitm_ctx,
-            firewall,
-            expected_message="catalog cache unavailable: cache_untrusted",
-            cache_mode=0o666,
-        )
-
-    def test_world_writable_runner_catalog_cache_reports_untrusted(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"fallback": _cache_firewall("fallback", "https://cache.example.com")},
-        )
-        cache_path.chmod(0o666)
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-fallback", "fallback")},
-        )
-
-        _assert_invalid_builtin_vm_with_cache(
-            registry_path=registry_path,
-            cache_path=cache_path,
-            mitm_ctx=mitm_ctx,
-            expected_message="catalog cache unavailable: cache_untrusted",
-        )
-
-    def test_malformed_aws_sigv4_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        del firewall["apis"][0]["auth"]["awsSigv4"]["secretAccessKey"]
-
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_host_policy_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["hostPolicy"] = {
-            "kind": "providerOwned",
-            "exactHosts": ["127.0.0.1"],
-        }
-
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_permission_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["permissions"][0]["rules"] = []
-
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_malformed_rule_runner_catalog_cache_fails_closed(self, tmp_path, mitm_ctx):
-        firewall = _cache_firewall("fallback", "https://cache.example.com")
-        firewall["apis"][0]["permissions"][0]["rules"] = ["GET /items/{path+}/tail"]
-
-        _assert_cache_firewall_is_invalid(tmp_path, mitm_ctx, firewall)
-
-    def test_runner_catalog_cache_change_invalidates_registry_snapshot(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": builtin_vm("run-mutable", "mutable")},
-        )
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"mutable": _cache_firewall("mutable", "https://cache-a.example.com")},
-        )
-        os.utime(cache_path, ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000))
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            _write_catalog_cache(
-                cache_path,
-                digest="sha256:bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                version="catalog-b",
-                firewalls={"mutable": _cache_firewall("mutable", "https://cache-b.example.com")},
-            )
-            os.utime(cache_path, ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001))
-            second_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, _ = first_context
-        second_vm_info, second_compiled, _ = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        first_api = first_vm_info["firewalls"][0]["apis"][0]
-        second_api = second_vm_info["firewalls"][0]["apis"][0]
-        assert first_api["base"] == "https://cache-a.example.com"
-        assert second_api["base"] == "https://cache-b.example.com"
-        first_runtime_policy = first_api[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER]
-        second_runtime_policy = second_api[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER]
-        assert isinstance(
-            first_runtime_policy,
-            builtin_host_policy.CompiledBuiltinHostPolicy,
-        )
-        assert isinstance(
-            second_runtime_policy,
-            builtin_host_policy.CompiledBuiltinHostPolicy,
-        )
-        assert first_runtime_policy is not second_runtime_policy
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-
-    def test_runner_catalog_cache_change_recompiles_core_when_metadata_is_unchanged(
-        self, tmp_path, mitm_ctx
-    ):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {
-                "10.200.0.1": {
-                    **builtin_vm("run-mutable", "mutable"),
-                    "networkPolicies": {
-                        "mutable": {
-                            "allow": ["read"],
-                            "deny": [],
-                            "unknownPolicy": "deny",
-                        }
-                    },
-                }
-            },
-        )
-        digest = "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-        version = "catalog-a"
-        _write_catalog_cache(
-            cache_path,
-            digest=digest,
-            version=version,
-            firewalls={
-                "mutable": _cache_firewall(
-                    "mutable",
-                    "https://cache.example.com",
-                    rules=["GET /items"],
-                )
-            },
-        )
-        os.utime(cache_path, ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000))
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            _write_catalog_cache(
-                cache_path,
-                digest=digest,
-                version=version,
-                firewalls={
-                    "mutable": _cache_firewall(
-                        "mutable",
-                        "https://cache.example.com",
-                        rules=["POST /items"],
-                    )
-                },
-            )
-            os.utime(cache_path, ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001))
-            second_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert first_context is not None
-        assert second_context is not None
-        _first_vm_info, first_compiled, first_policies = first_context
-        _second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-        first_result = matching.match_compiled_firewall_request(
-            "https://cache.example.com/items",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://cache.example.com/items",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallBlock)
-        assert second_result.reason == "unknown_endpoint"
-
-    def test_repeated_builtin_firewall_refs_share_core_but_keep_vm_api_ids(
-        self, tmp_path, mitm_ctx
-    ):
-        path, cache_path = _write_registry_with_cache(
-            tmp_path,
-            {
-                "10.200.0.1": builtin_vm("run-github-a", "github"),
-                "10.200.0.2": builtin_vm("run-github-b", "github"),
-            },
-            {"github": _github_cache_firewall()},
-        )
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(path))
-            second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is _first_firewall_core(second_compiled)
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://api.github.com/repos/vm0-ai/vm0",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://api.github.com/repos/vm0-ai/vm0",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_vm_info["firewalls"][0]["apis"][0]
-        assert second_result.api_entry is second_vm_info["firewalls"][0]["apis"][0]
-        assert first_result.api_entry["id"] == "run-github-a:0"
-        assert second_result.api_entry["id"] == "run-github-b:0"
-
-    def test_builtin_firewall_refs_share_static_payload_but_keep_vm_api_shells(
-        self, tmp_path, mitm_ctx
-    ):
-        auth = {"headers": {"Authorization": "Bearer ${{ secrets.API_TOKEN }}"}}
-        permissions = [{"name": "read-items", "rules": ["GET /items"]}]
-        path, cache_path = _write_registry_with_cache(
-            tmp_path,
-            {
-                "10.200.0.1": builtin_vm("run-large-a", "large"),
-                "10.200.0.2": builtin_vm("run-large-b", "large"),
-            },
-            {
-                "large": {
-                    "name": "large",
-                    "apis": [
-                        {
-                            "base": "https://api.example.com",
-                            "auth": auth,
-                            "permissions": permissions,
-                        },
-                        {
-                            "base": "https://upload.example.com",
-                            "auth": auth,
-                            "permissions": permissions,
-                        },
-                    ],
-                }
-            },
-        )
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(path))
-            second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is _first_firewall_core(second_compiled)
-
-        first_firewall = first_vm_info["firewalls"][0]
-        second_firewall = second_vm_info["firewalls"][0]
-        assert first_firewall is not second_firewall
-        assert first_firewall["apis"] is not second_firewall["apis"]
-
-        first_api = first_firewall["apis"][0]
-        second_api = second_firewall["apis"][0]
-        assert first_api is not second_api
-        assert first_api["id"] == "run-large-a:0"
-        assert second_api["id"] == "run-large-b:0"
-        assert first_api["permissions"] is second_api["permissions"]
-        assert first_api["auth"] is second_api["auth"]
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_api
-        assert second_result.api_entry is second_api
-
-    def test_builtin_refs_with_different_base_url_vars_do_not_share_core(self, tmp_path, mitm_ctx):
-        path, cache_path = _write_registry_with_cache(
-            tmp_path,
-            {
-                "10.200.0.1": builtin_vm(
-                    "run-zendesk-a",
-                    "zendesk",
-                    {"ZENDESK_SUBDOMAIN": "acme"},
-                ),
-                "10.200.0.2": builtin_vm(
-                    "run-zendesk-b",
-                    "zendesk",
-                    {"ZENDESK_SUBDOMAIN": "beta"},
-                ),
-            },
-            {"zendesk": _zendesk_cache_firewall()},
-        )
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(path))
-            second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-        assert first_vm_info["firewalls"][0]["apis"][0]["base"] == "https://acme.zendesk.com"
-        assert second_vm_info["firewalls"][0]["apis"][0]["base"] == "https://beta.zendesk.com"
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://acme.zendesk.com/api/v2/tickets",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://beta.zendesk.com/api/v2/tickets",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_vm_info["firewalls"][0]["apis"][0]
-        assert second_result.api_entry is second_vm_info["firewalls"][0]["apis"][0]
-
-    def test_builtin_compile_cache_is_scoped_to_registry_path(self, tmp_path, mitm_ctx):
-        path_a = tmp_path / "registry-a.json"
-        path_b = tmp_path / "registry-b.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        data = {"10.200.0.1": builtin_vm("run-github", "github")}
-        write_multi_vm_registry(path_a, data)
-        write_multi_vm_registry(path_b, data)
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"github": _github_cache_firewall()},
-        )
-
-        with mitm_ctx(
-            registry_path=str(path_a),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(path_a))
-        with mitm_ctx(
-            registry_path=str(path_b),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            second_context = registry.get_vm_context("10.200.0.1", str(path_b))
-
-        assert first_context is not None
-        assert second_context is not None
-        _, first_compiled, _ = first_context
-        _, second_compiled, _ = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-
-    def test_invalid_builtin_entry_does_not_poison_valid_vm_core_cache(self, tmp_path, mitm_ctx):
-        path, cache_path = _write_registry_with_cache(
-            tmp_path,
-            {
-                "10.200.0.1": builtin_vm("run-github", "github"),
-                "10.200.0.2": builtin_vm("run-missing", "missing-firewall"),
-            },
-            {"github": _github_cache_firewall()},
-        )
-
-        with (
-            mitm_ctx(
-                registry_path=str(path),
-                builtin_firewall_catalog_cache_path=str(cache_path),
-            ),
-            patch.object(registry.ctx, "log", MagicMock(), create=True),
-        ):
-            valid_context = registry.get_vm_context("10.200.0.1", str(path))
-            invalid_context = registry.get_vm_context("10.200.0.2", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert valid_context is not None
-        assert invalid_context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert state.invalid_vms["10.200.0.2"].reason == "invalid_firewalls"
-        _, compiled_firewalls, compiled_policies = valid_context
-        assert compiled_firewalls is not None
-        result = matching.match_compiled_firewall_request(
-            "https://api.github.com/repos/vm0-ai/vm0",
-            "GET",
-            compiled_firewalls,
-            compiled_policies,
-        )
-        assert isinstance(result, matching.FirewallAllow)
-
-    def test_builtin_core_cache_prunes_inactive_keys_on_registry_reload(self, tmp_path, mitm_ctx):
-        path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={
-                "cache-a": {
-                    "name": "cache-a",
-                    "apis": [
-                        {
-                            "base": "https://api-a.example.com",
-                            "auth": {"headers": {"Authorization": "Bearer token"}},
-                            "permissions": [{"name": "read", "rules": ["GET /items"]}],
-                        }
-                    ],
-                },
-                "cache-b": {
-                    "name": "cache-b",
-                    "apis": [
-                        {
-                            "base": "https://api-b.example.com",
-                            "auth": {"headers": {"Authorization": "Bearer token"}},
-                            "permissions": [{"name": "read", "rules": ["GET /items"]}],
-                        }
-                    ],
-                },
-            },
-        )
-        write_multi_vm_registry(path, {"10.200.0.1": builtin_vm("run-cache-a", "cache-a")})
-        os.utime(path, ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000))
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert first_context is not None
-        assert len(registry._registry_state.builtin_firewall_core_cache) == 1
-        first_cache_key = next(iter(registry._registry_state.builtin_firewall_core_cache))
-        assert first_cache_key[0] == "cache-a"
-
-        write_multi_vm_registry(path, {"10.200.0.1": builtin_vm("run-cache-b", "cache-b")})
-        os.utime(path, ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001))
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            second_context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert second_context is not None
-        assert len(registry._registry_state.builtin_firewall_core_cache) == 1
-        second_cache_key = next(iter(registry._registry_state.builtin_firewall_core_cache))
-        assert second_cache_key[0] == "cache-b"
-
-    def test_registry_unavailable_clears_compiled_core_but_retains_shared_catalog(
-        self, tmp_path, mitm_ctx
-    ):
-        path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"cache-a": _cache_firewall("cache-a", "https://api-a.example.com")},
-        )
-        write_multi_vm_registry(path, {"10.200.0.1": builtin_vm("run-cache-a", "cache-a")})
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            assert context is not None
-            assert len(registry._registry_state.builtin_firewall_core_cache) == 1
-            retained_catalog = builtin_firewall_cache._cache_state.catalog
-            assert retained_catalog is not None
-
-            path.write_text("{ broken")
-            unavailable = registry.load_registry_state(str(path))
-
-        assert isinstance(unavailable, registry.RegistryUnavailable)
-        assert unavailable.reason == "parse_failed"
-        assert registry._registry_state.builtin_firewall_core_cache == {}
-        assert builtin_firewall_cache._cache_state.catalog is retained_catalog
-
-    def test_registry_dropping_builtin_entries_retains_shared_catalog(self, tmp_path, mitm_ctx):
-        path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"cache-a": _cache_firewall("cache-a", "https://api-a.example.com")},
-        )
-        write_multi_vm_registry(path, {"10.200.0.1": builtin_vm("run-cache-a", "cache-a")})
-        os.utime(path, ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000))
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            assert context is not None
-            retained_catalog = builtin_firewall_cache._cache_state.catalog
-            assert retained_catalog is not None
-
-            write_multi_vm_registry(path, {"10.200.0.1": inline_vm("run-inline")})
-            os.utime(path, ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001))
-            inline_context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert inline_context is not None
-        assert builtin_firewall_cache._cache_state.catalog is retained_catalog
-
-    def test_no_builtin_registry_fast_path_retains_shared_catalog(self, tmp_path, mitm_ctx):
-        path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"cache-a": _cache_firewall("cache-a", "https://api-a.example.com")},
-        )
-        write_multi_vm_registry(path, {"10.200.0.1": inline_vm("run-inline")})
-
-        with mitm_ctx(
-            registry_path=str(path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(path))
-            assert first_context is not None
-            assert builtin_firewall_cache._cache_state.catalog is None
-
-            snapshot = registry_firewalls.load_catalog_snapshot(str(cache_path))
-            assert snapshot.catalog is not None
-            assert snapshot.dependency_file_key is not None
-            assert snapshot.dependency_file_key.absolute_path == str(cache_path.absolute())
-            assert snapshot.catalog.identity.source == "cache"
-            assert snapshot.catalog.identity.catalog_version == "catalog-a"
-            assert snapshot.catalog.identity.file_key == snapshot.dependency_file_key
-            retained_catalog = builtin_firewall_cache._cache_state.catalog
-            assert retained_catalog is snapshot.catalog
-
-            second_context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert second_context is not None
-        assert builtin_firewall_cache._cache_state.catalog is retained_catalog
-
-    def test_inline_firewalls_do_not_share_compiled_core(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_multi_vm_registry(
-            path,
-            {
-                "10.200.0.1": inline_vm("run-inline-a"),
-                "10.200.0.2": inline_vm("run-inline-b"),
-            },
-        )
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_vm_info["firewalls"][0]["apis"][0]
-        assert second_result.api_entry is second_vm_info["firewalls"][0]["apis"][0]
-
-    def test_inline_only_registry_ignores_catalog_cache_changes(self, tmp_path, mitm_ctx):
-        registry_path = tmp_path / "registry.json"
-        cache_path = tmp_path / "builtin-firewall-catalog-cache.json"
-        write_multi_vm_registry(
-            registry_path,
-            {"10.200.0.1": inline_vm("run-inline")},
-        )
-        _write_catalog_cache(
-            cache_path,
-            digest="sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            version="catalog-a",
-            firewalls={"unused": _cache_firewall("unused", "https://unused-a.example.com")},
-        )
-        os.utime(cache_path, ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000))
-
-        with mitm_ctx(
-            registry_path=str(registry_path),
-            builtin_firewall_catalog_cache_path=str(cache_path),
-        ):
-            first_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-            _write_catalog_cache(
-                cache_path,
-                digest="sha256:bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                version="catalog-b",
-                firewalls={"unused": _cache_firewall("unused", "https://unused-b.example.com")},
-            )
-            os.utime(cache_path, ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001))
-            second_context = registry.get_vm_context("10.200.0.1", str(registry_path))
-
-        assert first_context is not None
-        assert second_context is not None
-        _first_vm_info, first_compiled, _first_policies = first_context
-        _second_vm_info, second_compiled, _second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is _first_firewall_core(second_compiled)
-
-    def test_inline_firewall_api_ids_preserve_custom_ids_and_global_positions(self, tmp_path):
-        path = tmp_path / "registry.json"
-        vm = inline_vm("run-inline")
-        first_api = vm["firewalls"][0]["firewall"]["apis"][0]
-        first_api["id"] = "custom-api-id"
-        vm["firewalls"].append(
-            {
-                "kind": "inline",
-                "firewall": {
-                    "name": "upload",
-                    "apis": [{**first_api, "id": "", "base": "https://upload.example.com"}],
-                },
-            }
-        )
-        write_multi_vm_registry(path, {"10.200.0.1": vm})
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert [firewall["apis"][0]["id"] for firewall in vm_info["firewalls"]] == [
-            "custom-api-id",
-            "run-inline:1",
-        ]
